+++ text-compression/src/test/java/com/giljulio/compression/text/CrunchTest.java
@@ -23,7 +23,7 @@ public class CrunchTest {
     public void when_uncompressCharactersAndReferences_validateOutput() {
         ReaderItem[] items = new ReaderItem[]{
                 new Character('a'), new Character('b'), new Character('c'),
-                new Character('d'), new Character('e'), new Reference(-5, 3)};
+                new Character('d'), new Character('e'), new Reference((short) -5, 3)};
         Crunch underTest = new Crunch.Builder().build();
 
         String output = underTest.decompress(new MockDecompressReader(items));
